+++ server/hazard-service/src/main/java/edu/illinois/ncsa/incore/service/hazard/controllers/HurricaneController.java
@@ -5,6 +5,7 @@ import com.fasterxml.jackson.core.type.TypeReference;
 import com.fasterxml.jackson.databind.ObjectMapper;
 import edu.illinois.ncsa.incore.common.AllocationConstants;
 import edu.illinois.ncsa.incore.common.HazardConstants;
+import edu.illinois.ncsa.incore.common.auth.Authorizer;
 import edu.illinois.ncsa.incore.common.auth.IAuthorizer;
 import edu.illinois.ncsa.incore.common.auth.Privileges;
 import edu.illinois.ncsa.incore.common.dao.ICommonRepository;
@@ -260,6 +261,7 @@ public class HurricaneController {
                     }
 
                     hurricane.setCreator(this.username);
+                    hurricane.setOwner(this.username);
                     hurricane = repository.addHurricane(hurricane);
 
                     Space space = spaceRepository.getSpaceByName(this.username);
@@ -396,7 +398,8 @@ public class HurricaneController {
     public Hurricane deleteHurricanes(@Parameter(name = "Hurricane Id", required = true) @PathParam("hurricane-id") String hurricaneId) {
         Hurricane hurricane = getHurricaneById(hurricaneId);
 
-        if (authorizer.canUserDeleteMember(this.username, hurricaneId, spaceRepository.getAllSpaces(), this.groups)) {
+        Boolean isAdmin = Authorizer.getInstance().isUserAdmin(this.groups);
+        if (this.username.equals(hurricane.getOwner()) || isAdmin) {
             // delete associated datasets
             if (hurricane != null && hurricane instanceof HurricaneDataset) {
                 HurricaneDataset hurrDataset = (HurricaneDataset) hurricane;